Add chap02 sec09 part01 EmptyFrameViewer program
A graphical application shows drawings, or information, inside of a
frame: a window with a title bar.
To display a frame in Java we can use the JFrame class of the
javax.swing package in the Java library. Swing is the nickname for the
graphical user interface library in Java. The x denotes the fact that
swing package started out as a Java extension before it was added to
the standard library.