Add chap02 sec09 Rectangle Component program
You cannot draw directly onto a frame. Drawing happens in a Component
object. In the swing toolkit, the JComponent class represents a blank
component.
Becuase we don't want to add a blank component we have to modify the
JComponent class and specify how the component should be painted.
The solution is to declare a new class that extends the JComponent
class.
The extends reserved word indicates that our component class can be
used like JComponent. However our component class will be
different from the JComponent class in one respect: its paintComponent
method will contain instructions to draw something.
When the component is shown for the first time the paintComponent
method is called automatically.