Add ch07 proj09 3InARow method and test-suite
This commit implements a method for checking if a row
in the grid contains 3 identical elements in a row.

The tic-tac-toe game is still missing implementation
for checking if columns or diagonals have identical
elements. The public interface of those methdos is
there, but their implementation still needs to filled
in.

This commit also adds a test suite folder with a test
input sequence which should end with a 3 in a row win
for the circle player.
There is a batch file in the test suite folder which
contains a batch of commands to be executed by the
operating system through the shell.