Add/fix methods & tests for royalFlush/cardCount methods
This commit adds a method which checks if a hand is a royal flush.
This commit also fixes the method which count the most common
card value in a hand of poker. Which determines whether a hand
has a single pair, three of a kind, or four of a kind.
Finally, this commit adds tests for both methods previously
mentioned which test if the royal flush and most common card
value counting methods work.