Update public interface of game of life class
This commit updates the public interface of the game of life class
to account for a game with a board which increases in size.
This way game of life configurations like the glider gun can be
implemented in this game, and results like the example glider gun
in the book can be obtained by playing.