Update game to use 1 scanner for testing purposes
This commit will add a new implementation to the game
of life which uses a single scanner class in order to
make the expected behaviour of the class easier to
predict when redirecting input from external text
files for testing purposes.

This commit also makes the game of life tests
re-usable with different input text files in order to
allow easier expansion of test inputs.

bug notice:
 - The shell script containing a batch of commands
   which compile and run all of the tests does not
   work with the provided test input text files and
   tester class implementations.