Fix tests input redirection from txt files
This commit fixes the errors with input redirection
from text files when running tests on the game of
life player with sample input. This way tests run,
and now the expected result of a test is formatted
properly and present in the output text file where
output is redirected.

bugs:
 - thanks to being able to run tests properly it
   was discovered that some of the tests to not
   pass because there is an error with the
   creation of board generations sometimes.