Correct glider gun test starting configuration
This commit corrects the starting configuration of
the game of life board in the glider gun test
input file.

The actual result is now the same as the expected
result, and the game of life works as expected.

The mistake was the the 4-cell square of cells
at the top of the starting configuration were
mistakenly 1 column of cells to the left compared
to were they were meant to be.